# runtime/src/enclave_rpc/client: Support concurrent sessions