--- a/.changelog/5879.feature.md
+++ b/.changelog/5879.feature.md
@@ -1,0 +1,1 @@
+runtime/src/enclave_rpc/client: Support concurrent sessions