--- a/runtime/src/enclave_rpc/mod.rs
+++ b/runtime/src/enclave_rpc/mod.rs
@@ -5,6 +5,7 @@ pub mod context;
 pub mod demux;
 pub mod dispatcher;
 pub mod session;
+pub mod sessions;
 mod transport;
 pub mod types;