Measured locally execution time on `compute-0` of `loop` in method `churp_recover_state_key` in test `e2e.runtime.keymanager-churp-txs`. 

Speed-up looks lower than expected, but I didn't investigate the numbers.

```
Before:
Time elapsed: 31.48498ms
Time elapsed: 33.393901ms
Time elapsed: 32.392303ms
Time elapsed: 32.062076ms

After:
Time elapsed: 22.728806ms
Time elapsed: 27.323837ms
Time elapsed: 22.350006ms
Time elapsed: 23.219776ms
```